+++ sulong/tests/com.oracle.truffle.llvm.tests.linker.native/dynLink/shadowing/variable/main.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2020, Oracle and/or its affiliates.
+ * Copyright (c) 2020, Oracle and/or its affiliates.
  *
  * All rights reserved.
  *
@@ -28,15 +28,13 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-void __clear_cache(void *begin, void *end) {
-  /*
-   * Old llvm/gcc+dragonegg versions do not intrinsify __builtin___clear_cache but replace it with a call to
-   * __clear_cache from libgcc. Lets make sure the target exist even if not linked against libgcc.
-   */
+#include <stdio.h>
+
+char *hello() {
+  return "function";
 }
 
 int main() {
-  char a[100];
-  __builtin___clear_cache(a, a + 100);
+  printf("Hello %s\n", hello());
   return 0;
 }